Write testList

+++ Занятие 7 (Исключения, Файлы)/Task Management Service/src/test/java/TaskManagenentServiceTest.java
@@ -11,18 +11,17 @@ import java.io.File;
 import java.io.IOException;
 
 public class TaskManagenentServiceTest {
-    // TODO: manage, process**
-    // TODO: findTask/Executor с разными по типу id
+
+    static String testStoragePath = System.getProperty("user.dir") + "\\TaskManagenentServiceTest\\";
+    static TaskManagementService taskService = new TaskManagementService(System.in, testStoragePath);
 
     @AfterEach
-    public void flushCounts() {
+    public void clear() {
+        deleteDirectory(new File(testStoragePath), false);
         CountSaver emptyCountSaver = new CountSaver("Folder1", "Folder2");
         Executor.loadCount(emptyCountSaver);
         Task.loadCount(emptyCountSaver);
     }
-
-    static String testStoragePath = System.getProperty("user.dir") + "\\TaskManagenentServiceTest\\";
-    static TaskManagementService taskService = new TaskManagementService(System.in, testStoragePath);
 
     @Test
     public void testWrongCommands() {
@@ -51,32 +50,63 @@ public class TaskManagenentServiceTest {
     public void testAddTaskWithoutExecutor() throws ClassNotFoundException, IOException, NoTaskException {
         taskService.setIn(new ByteArrayInputStream("Big Problem\n\nSome description".getBytes()));
         Assertions.assertEquals("Task added and saved", taskService.manage("add -t"));
-        Task task = taskService.findTask("Task-1");
+        Task task = taskService.findTask("1");
         Assertions.assertEquals("Big Problem", task.getName());
         Assertions.assertEquals("Some description", task.getDescription());
         Assertions.assertEquals(TaskStatus.TODO, task.getStatus());
         Assertions.assertNull(task.getExecutor());
     }
 
-    @Test
-    public void testAddTaskWithExecutor() {
+    private void addExecutor(String name) {
+        taskService.setIn(new ByteArrayInputStream(name.getBytes()));
+        Assertions.assertEquals("Executor added and saved", taskService.manage("add -e"));
+    }
 
+    private void addTask(String name, String executorId) {
+        taskService.setIn(new ByteArrayInputStream((name + "\n" + executorId + "\n ").getBytes()));
+        Assertions.assertEquals("Task added and saved", taskService.manage("add -t"));
+    }
+
+    @Test
+    public void testAddTaskWithExecutor() throws ClassNotFoundException, IOException, NoTaskException, NoExecutorException {
+        addExecutor("Georgy");
+        Executor executor = taskService.findExecutor("1");
+        taskService.setIn(new ByteArrayInputStream("Big Problem\n1\nSome description".getBytes()));
+        Assertions.assertEquals("Task added and saved", taskService.manage("add -t"));
+        Task task = taskService.findTask("1");
+        Assertions.assertEquals(executor, task.getExecutor());
+    }
+
+    @Test
+    public void testList() {
+        addExecutor("Georgy");
+        addExecutor("Alex");
+        addExecutor("Anton");
+        addTask("Some Problem", "");
+        addTask("Big Problem", "1");
+        Assertions.assertTrue(taskService.manage("list").matches(
+                "\\tTasks:\\nTask.*\\nTask.*\\n\\tExecutors:\\nExecutor.*\\nExecutor.*\\nExecutor.*\\n"
+        ));
     }
 
 
 
     @AfterAll
-    static void clear() {
-        deleteDirectory(new File(testStoragePath));
+    static void delete() {
+        deleteDirectory(new File(testStoragePath), true);
     }
 
-    private static void deleteDirectory(File directory) {
+    private static void deleteDirectory(File directory, boolean deldir) {
         File[] allContents = directory.listFiles();
         if (allContents != null) {
             for (File file : allContents) {
-                deleteDirectory(file);
+                deleteDirectory(file, deldir);
             }
         }
-        directory.delete();
+        if (deldir)
+            directory.delete();
+        else
+            if (directory.isFile())
+                directory.delete();
     }
 }
